🐛 move fetch()
