+++ app/lib/data.ts
@@ -7,22 +7,13 @@ import { sql } from '@vercel/postgres';
 import camelcaseKeys from "camelcase-keys";
 import decamelizeKeys from "decamelize-keys";
 
-export async function fetchPlayers() {
-    try {
-        const data = await sql<Player>`SELECT * FROM players;`;
-        return camelcaseKeys(data.rows);
-    } catch (error) {
-        console.error('Database Error:', error);
-        throw new Error('Failed to fetch the players.');
-    }
-}
 
 export async function fetchParticipants() {
     try {
         const data = await sql<Player>`
-            SELECT * FROM players
-            WHERE is_player = TRUE
-            ORDER BY random();
+        SELECT * FROM players
+        WHERE is_player = TRUE
+        ORDER BY random();
         `;
         return camelcaseKeys(data.rows);
     } catch (error) {
